requirements.txt
astroid==2.4.2
autopep8==1.5.4
boto==2.49.0
boto3==1.16.4
botocore==1.19.4
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
idna==2.10
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.11
wrapt==1.12.1
